# backend/admin/docs/swagger.json
{
    "basePath": "/",
    "definitions": {
...
        "http.ErrorResponse": {
            "additionalProperties": false,
            "example": {
                "error": "error message"
            },
            "properties": {
                "error": {
                    "example": "error message",
                    "type": "string"
                }
            },
            "type": "object"
        },
...
        "search.AutocompleteResponse": {
            "additionalProperties": false,
            "properties": {
                "values": {
                    "items": {
                        "type": "string"
                    },
                    "type": "array"
                }
            },
            "type": "object"
        },
...
    },
    "info": {
        "contact": {},
        "description": "API for Furocoumarins admin backend",
        "title": "Furocoumarins Admin API",
        "version": "1.0"
    },
    "paths": {
...
        "/autocomplete/{column}": {
            "get": {
                "description": "Returns prefix-matched values for a column",
                "parameters": [
                    {
                        "description": "Column name",
                        "in": "path",
                        "name": "column",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Prefix to match",
                        "in": "query",
                        "name": "value",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/search.AutocompleteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                },
                "summary": "Autocomplete column values",
                "tags": [
                    "search"
                ]
            }
        },
...
        "/make-table-active/{timestamp}": {
            "post": {
                "description": "Sets the given table as the active one",
                "parameters": [
                    {
                        "description": "Table timestamp",
                        "in": "path",
                        "name": "timestamp",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Activate table by timestamp",
                "tags": [
                    "tables"
                ]
            }
        },
...
        "/pages/{name}": {
            "get": {
                "description": "Returns markdown content of a page from S3",
                "parameters": [
                    {
                        "description": "Page name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "text/markdown"
                ],
                "responses": {
                    "200": {
                        "description": "Markdown content",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                },
                "summary": "Get page content by name",
                "tags": [
                    "pages"
                ]
            },
            "put": {
                "consumes": [
                    "application/octet-stream"
                ],
                "description": "Uploads markdown content for a page (admin only)",
                "parameters": [
                    {
                        "description": "Page name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "description": "Markdown content",
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Create or update page",
                "tags": [
                    "pages"
                ]
            }
        },
...
        "/table/{timestamp}": {
            "delete": {
                "description": "Deletes the table with the given timestamp",
                "parameters": [
                    {
                        "description": "Table timestamp",
                        "in": "path",
                        "name": "timestamp",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Delete table by timestamp",
                "tags": [
                    "tables"
                ]
            }
        },
        "/tables": {
            "delete": {
                "description": "Deletes all tables that are not marked as OK",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Delete all bad tables",
                "tags": [
                    "tables"
                ]
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "in": "header",
            "name": "Authorization",
            "type": "apiKey"
        }
    },
    "swagger": "2.0"
}